requirements.txt
ctransformers==0.2.5
pinecone-client
langchain
sentence-transformers==3.2.1
flask
pypdf
python-dotenv
-e .